+++ crates/rust-client/src/tests.rs
@@ -62,8 +62,8 @@ use crate::{
         mock::{MockClient, MockRpcApi},
     },
     transaction::{
-        DiscardCause, PaymentTransactionData, TransactionRequestBuilder, TransactionRequestError,
-        TransactionStatus,
+        DiscardCause, PaymentNoteDescription, SwapTransactionData, TransactionRequestBuilder,
+        TransactionRequestError, TransactionStatus,
     },
     utils::{execute_tx, execute_tx_and_sync, setup_accounts_and_faucets},
 };
@@ -92,19 +92,20 @@ pub async fn create_test_client_builder() -> (ClientBuilder, MockRpcApi, Filesys
     let arc_rpc_api = Arc::new(rpc_api.clone());
 
     let builder = ClientBuilder::new()
-        .with_rpc(arc_rpc_api)
-        .with_rng(Box::new(rng))
-        .with_store(store)
-        .with_filesystem_keystore(keystore_path.to_str().unwrap())
+        .rpc(arc_rpc_api)
+        .rng(Box::new(rng))
+        .store(store)
+        .filesystem_keystore(keystore_path.to_str().unwrap())
         .in_debug_mode(true)
-        .with_tx_graceful_blocks(None);
+        .tx_graceful_blocks(None);
 
     (builder, rpc_api, keystore)
 }
 
 pub async fn create_test_client() -> (MockClient, MockRpcApi, FilesystemKeyStore<StdRng>) {
     let (builder, rpc_api, keystore) = create_test_client_builder().await;
-    let client = builder.build().await.unwrap();
+    let mut client = builder.build().await.unwrap();
+    client.ensure_genesis_in_place().await.unwrap();
 
     (client, rpc_api, keystore)
 }
@@ -637,7 +638,7 @@ async fn transaction_request_expiration() {
             .unwrap();
 
     let transaction_request = TransactionRequestBuilder::new()
-        .with_expiration_delta(5)
+        .expiration_delta(5)
         .build_mint_fungible_asset(
             FungibleAsset::new(faucet.id(), 5u64).unwrap(),
             AccountId::try_from(ACCOUNT_ID_REGULAR_PUBLIC_ACCOUNT_IMMUTABLE_CODE).unwrap(),
@@ -688,7 +689,7 @@ async fn import_processing_note_returns_error() {
 
     let input = [(note.try_into().unwrap(), None)];
     let consume_note_request = TransactionRequestBuilder::new()
-        .with_unauthenticated_input_notes(input)
+        .unauthenticated_input_notes(input)
         .build()
         .unwrap();
     let transaction = client
@@ -736,7 +737,7 @@ async fn note_without_asset() {
 
     // Create and execute transaction
     let transaction_request = TransactionRequestBuilder::new()
-        .with_own_output_notes(vec![OutputNote::Full(note)])
+        .own_output_notes(vec![OutputNote::Full(note)])
         .build()
         .unwrap();
 
@@ -751,7 +752,7 @@ async fn note_without_asset() {
     let note = Note::new(vault, metadata, recipient);
 
     let transaction_request = TransactionRequestBuilder::new()
-        .with_own_output_notes(vec![OutputNote::Full(note)])
+        .own_output_notes(vec![OutputNote::Full(note)])
         .build()
         .unwrap();
 
@@ -766,8 +767,7 @@ async fn note_without_asset() {
 
     let error = TransactionRequestBuilder::new()
         .build_pay_to_id(
-            PaymentTransactionData::new(vec![], faucet.id(), wallet.id()),
-            None,
+            PaymentNoteDescription::new(vec![], faucet.id(), wallet.id()),
             NoteType::Public,
             client.rng(),
         )
@@ -777,12 +777,11 @@ async fn note_without_asset() {
 
     let error = TransactionRequestBuilder::new()
         .build_pay_to_id(
-            PaymentTransactionData::new(
+            PaymentNoteDescription::new(
                 vec![Asset::Fungible(FungibleAsset::new(faucet.id(), 0).unwrap())],
                 faucet.id(),
                 wallet.id(),
             ),
-            None,
             NoteType::Public,
             client.rng(),
         )
@@ -878,7 +877,6 @@ async fn real_note_roundtrip() {
 #[tokio::test]
 async fn added_notes() {
     let (mut client, _, authenticator) = create_test_client().await;
-    wait_for_node(&mut client).await;
 
     let faucet_account_header =
         insert_new_fungible_faucet(&mut client, AccountStorageMode::Private, &authenticator)
@@ -911,7 +909,6 @@ async fn added_notes() {
 #[tokio::test]
 async fn p2id_transfer() {
     let (mut client, _, authenticator) = create_test_client().await;
-    wait_for_node(&mut client).await;
 
     let (wallets, faucets) = setup_accounts_and_faucets(
         &mut client,
@@ -935,12 +932,11 @@ async fn p2id_transfer() {
     println!("Running P2ID tx...");
     let tx_request = TransactionRequestBuilder::new()
         .build_pay_to_id(
-            PaymentTransactionData::new(
+            PaymentNoteDescription::new(
                 vec![Asset::Fungible(asset)],
                 from_account_id,
                 to_account_id,
             ),
-            None,
             NoteType::Private,
             client.rng(),
         )
@@ -1019,7 +1015,6 @@ async fn p2id_transfer() {
 #[tokio::test]
 async fn p2id_transfer_failing_not_enough_balance() {
     let (mut client, _, authenticator) = create_test_client().await;
-    wait_for_node(&mut client).await;
 
     let (wallets, faucets) = setup_accounts_and_faucets(
         &mut client,
@@ -1041,12 +1036,11 @@ async fn p2id_transfer_failing_not_enough_balance() {
     println!("Running P2ID tx...");
     let tx_request = TransactionRequestBuilder::new()
         .build_pay_to_id(
-            PaymentTransactionData::new(
+            PaymentNoteDescription::new(
                 vec![Asset::Fungible(asset)],
                 from_account_id,
                 to_account_id,
             ),
-            None,
             NoteType::Private,
             client.rng(),
         )
@@ -1064,9 +1058,8 @@ async fn p2id_transfer_failing_not_enough_balance() {
 }
 
 #[tokio::test]
-async fn p2idr_transfer_consumed_by_target() {
+async fn p2ide_transfer_consumed_by_target() {
     let (mut client, _, authenticator) = create_test_client().await;
-    wait_for_node(&mut client).await;
 
     let (wallets, faucets) = setup_accounts_and_faucets(
         &mut client,
@@ -1132,15 +1125,15 @@ async fn p2idr_transfer_consumed_by_target() {
         .unwrap_or(0);
     let current_block_num = client.get_sync_height().await.unwrap();
     let asset = FungibleAsset::new(faucet_account_id, TRANSFER_AMOUNT).unwrap();
-    println!("Running P2IDR tx...");
+    println!("Running P2IDE tx...");
     let tx_request = TransactionRequestBuilder::new()
         .build_pay_to_id(
-            PaymentTransactionData::new(
+            PaymentNoteDescription::new(
                 vec![Asset::Fungible(asset)],
                 from_account_id,
                 to_account_id,
-            ),
-            Some(current_block_num + RECALL_HEIGHT_DELTA),
+            )
+            .with_reclaim_height(current_block_num + RECALL_HEIGHT_DELTA),
             NoteType::Private,
             client.rng(),
         )
@@ -1154,7 +1147,7 @@ async fn p2idr_transfer_consumed_by_target() {
     let notes = client.get_input_notes(NoteFilter::Committed).await.unwrap();
     assert!(!notes.is_empty());
 
-    // Make the `to_account_id` consume P2IDR note
+    // Make the `to_account_id` consume P2IDE note
     let note_id = tx_request.expected_output_own_notes().pop().unwrap().id();
     println!("Consuming Note...");
     let tx_request = TransactionRequestBuilder::new().build_consume_notes(vec![note_id]).unwrap();
@@ -1187,9 +1180,8 @@ async fn p2idr_transfer_consumed_by_target() {
 }
 
 #[tokio::test]
-async fn p2idr_transfer_consumed_by_sender() {
+async fn p2ide_transfer_consumed_by_sender() {
     let (mut client, mock_rpc_api, authenticator) = create_test_client().await;
-    wait_for_node(&mut client).await;
 
     let (wallets, faucets) = setup_accounts_and_faucets(
         &mut client,
@@ -1219,15 +1211,15 @@ async fn p2idr_transfer_consumed_by_sender() {
         .unwrap_or(0);
     let current_block_num = client.get_sync_height().await.unwrap();
     let asset = FungibleAsset::new(faucet_account_id, TRANSFER_AMOUNT).unwrap();
-    println!("Running P2IDR tx...");
+    println!("Running P2IDE tx...");
     let tx_request = TransactionRequestBuilder::new()
         .build_pay_to_id(
-            PaymentTransactionData::new(
+            PaymentNoteDescription::new(
                 vec![Asset::Fungible(asset)],
                 from_account_id,
                 to_account_id,
-            ),
-            Some(current_block_num + RECALL_HEIGHT_DELTA),
+            )
+            .with_reclaim_height(current_block_num + RECALL_HEIGHT_DELTA),
             NoteType::Private,
             client.rng(),
         )
@@ -1287,6 +1279,74 @@ async fn p2idr_transfer_consumed_by_sender() {
 }
 
 #[tokio::test]
+async fn p2ide_timelocked() {
+    let (mut client, mock_rpc_api, authenticator) = create_test_client().await;
+
+    let (wallets, faucets) = setup_accounts_and_faucets(
+        &mut client,
+        &authenticator,
+        AccountStorageMode::Private,
+        2,
+        1,
+        vec![vec![MINT_AMOUNT, 0]],
+    )
+    .await
+    .unwrap();
+
+    let from_account_id = wallets[0].id();
+    let to_account_id = wallets[1].id();
+    let faucet_account_id = faucets[0].id();
+
+    let current_block_num = client.get_sync_height().await.unwrap();
+
+    let asset = FungibleAsset::new(faucet_account_id, TRANSFER_AMOUNT).unwrap();
+    let tx_request = TransactionRequestBuilder::new()
+        .build_pay_to_id(
+            PaymentNoteDescription::new(
+                vec![Asset::Fungible(asset)],
+                from_account_id,
+                to_account_id,
+            )
+            .with_timelock_height(current_block_num + RECALL_HEIGHT_DELTA)
+            .with_reclaim_height(current_block_num),
+            NoteType::Public,
+            client.rng(),
+        )
+        .unwrap();
+    let note = tx_request.expected_output_own_notes().pop().unwrap();
+
+    execute_tx_and_sync(&mut client, from_account_id, tx_request).await.unwrap();
+
+    // Check that it's still too early to consume by both accounts
+    let tx_request = TransactionRequestBuilder::new().build_consume_notes(vec![note.id()]).unwrap();
+    let results = [
+        client.new_transaction(from_account_id, tx_request.clone()).await,
+        client.new_transaction(to_account_id, tx_request).await,
+    ];
+    assert!(results.iter().all(|result| {
+        result.as_ref().is_err_and(|err| {
+            matches!(
+                err,
+                ClientError::TransactionExecutorError(
+                    TransactionExecutorError::TransactionProgramExecutionFailed(_)
+                )
+            )
+        })
+    }));
+
+    // Wait to consume with the target account
+    mock_rpc_api.advance_blocks(RECALL_HEIGHT_DELTA);
+    client.sync_state().await.unwrap();
+
+    // Consume the note with the target account
+    let tx_request = TransactionRequestBuilder::new().build_consume_notes(vec![note.id()]).unwrap();
+    execute_tx_and_sync(&mut client, to_account_id, tx_request).await.unwrap();
+
+    let target_account: Account = client.get_account(to_account_id).await.unwrap().unwrap().into();
+    assert_eq!(target_account.vault().get_balance(faucet_account_id).unwrap(), TRANSFER_AMOUNT);
+}
+
+#[tokio::test]
 async fn get_consumable_notes() {
     let (mut client, _, authenticator) = create_test_client().await;
 
@@ -1323,15 +1383,15 @@ async fn get_consumable_notes() {
 
     // Do a transfer from first account to second account
     let asset = FungibleAsset::new(faucet_account_id, TRANSFER_AMOUNT).unwrap();
-    println!("Running P2IDR tx...");
+    println!("Running P2IDE tx...");
     let tx_request = TransactionRequestBuilder::new()
         .build_pay_to_id(
-            PaymentTransactionData::new(
+            PaymentNoteDescription::new(
                 vec![Asset::Fungible(asset)],
                 from_account_id,
                 to_account_id,
-            ),
-            Some(100.into()),
+            )
+            .with_reclaim_height(100.into()),
             NoteType::Private,
             client.rng(),
         )
@@ -1402,12 +1462,11 @@ async fn get_output_notes() {
     println!("Running P2ID tx...");
     let tx_request = TransactionRequestBuilder::new()
         .build_pay_to_id(
-            PaymentTransactionData::new(
+            PaymentNoteDescription::new(
                 vec![Asset::Fungible(asset)],
                 from_account_id,
                 random_account_id,
             ),
-            None,
             NoteType::Private,
             client.rng(),
         )
@@ -1429,8 +1488,7 @@ async fn get_output_notes() {
 async fn account_rollback() {
     let (builder, rpc_api, authenticator) = create_test_client_builder().await;
 
-    let mut client =
-        builder.with_tx_graceful_blocks(Some(TX_GRACEFUL_BLOCKS)).build().await.unwrap();
+    let mut client = builder.tx_graceful_blocks(Some(TX_GRACEFUL_BLOCKS)).build().await.unwrap();
 
     client.sync_state().await.unwrap();
 
@@ -1445,8 +1503,7 @@ async fn account_rollback() {
 
     let tx_request = TransactionRequestBuilder::new()
         .build_pay_to_id(
-            PaymentTransactionData::new(vec![Asset::Fungible(asset)], account_id, account_id),
-            None,
+            PaymentNoteDescription::new(vec![Asset::Fungible(asset)], account_id, account_id),
             NoteType::Public,
             client.rng(),
         )
@@ -1527,10 +1584,9 @@ async fn subsequent_discarded_transactions() {
     // Create a transaction that will expire in 2 blocks
     let asset = FungibleAsset::new(faucet_account_id, TRANSFER_AMOUNT).unwrap();
     let tx_request = TransactionRequestBuilder::new()
-        .with_expiration_delta(2)
+        .expiration_delta(2)
         .build_pay_to_id(
-            PaymentTransactionData::new(vec![Asset::Fungible(asset)], account_id, account_id),
-            None,
+            PaymentNoteDescription::new(vec![Asset::Fungible(asset)], account_id, account_id),
             NoteType::Public,
             client.rng(),
         )
@@ -1549,8 +1605,7 @@ async fn subsequent_discarded_transactions() {
     let asset = FungibleAsset::new(faucet_account_id, TRANSFER_AMOUNT).unwrap();
     let tx_request = TransactionRequestBuilder::new()
         .build_pay_to_id(
-            PaymentTransactionData::new(vec![Asset::Fungible(asset)], account_id, account_id),
-            None,
+            PaymentNoteDescription::new(vec![Asset::Fungible(asset)], account_id, account_id),
             NoteType::Public,
             client.rng(),
         )
@@ -1621,7 +1676,7 @@ async fn missing_recipient_digest() {
     let dummy_recipient_digest = dummy_recipient.digest();
 
     let tx_request = TransactionRequestBuilder::new()
-        .with_expected_output_recipients(vec![dummy_recipient])
+        .expected_output_recipients(vec![dummy_recipient])
         .build_mint_fungible_asset(
             FungibleAsset::new(faucet.id(), 5u64).unwrap(),
             AccountId::try_from(ACCOUNT_ID_PRIVATE_SENDER).unwrap(),
@@ -1635,4 +1690,164 @@ async fn missing_recipient_digest() {
     if let ClientError::MissingOutputRecipients(digests) = error {
         assert!(digests == vec![dummy_recipient_digest]);
     }
+}
+
+#[tokio::test]
+async fn input_note_checks() {
+    let (mut client, _, authenticator) = create_test_client().await;
+
+    let (wallet, faucet) =
+        setup_wallet_and_faucet(&mut client, AccountStorageMode::Private, &authenticator).await;
+
+    let mut mint_notes = vec![];
+    for _ in 0..10 {
+        mint_notes.push(mint_note(&mut client, wallet.id(), faucet.id(), NoteType::Public).await);
+    }
+
+    let duplicate_note_tx_request = TransactionRequestBuilder::new()
+        .build_consume_notes(vec![mint_notes[0].id(), mint_notes[0].id()]);
+
+    assert!(matches!(
+        duplicate_note_tx_request,
+        Err(TransactionRequestError::DuplicateInputNote(note_id)) if note_id == mint_notes[0].id()
+    ));
+
+    let tx_request = TransactionRequestBuilder::new()
+        .build_consume_notes(mint_notes.iter().map(InputNoteRecord::id).collect())
+        .unwrap();
+
+    let transaction = client.new_transaction(wallet.id(), tx_request).await.unwrap();
+
+    let input_notes = transaction.executed_transaction().input_notes().iter();
+
+    // Check that the input notes have the same order as the original notes
+    for (i, input_note) in input_notes.enumerate() {
+        assert_eq!(input_note.id(), mint_notes[i].id());
+    }
+
+    let tx_id = transaction.executed_transaction().id();
+    client.submit_transaction(transaction).await.unwrap();
+    wait_for_tx(&mut client, tx_id).await;
+
+    // Check that using consumed notes will return an error
+    let consumed_note_tx_request = TransactionRequestBuilder::new()
+        .build_consume_notes(vec![mint_notes[0].id()])
+        .unwrap();
+    let error = client.new_transaction(wallet.id(), consumed_note_tx_request).await.unwrap_err();
+
+    assert!(matches!(
+        error,
+        ClientError::TransactionRequestError(TransactionRequestError::InputNoteAlreadyConsumed(_))
+    ));
+
+    // Check that adding an authenticated note that is not tracked by the client will return an
+    // error
+    let missing_authenticated_note_tx_request = TransactionRequestBuilder::new()
+        .build_consume_notes(vec![EMPTY_WORD.into()])
+        .unwrap();
+    let error = client
+        .new_transaction(wallet.id(), missing_authenticated_note_tx_request)
+        .await
+        .unwrap_err();
+
+    assert!(matches!(
+        error,
+        ClientError::TransactionRequestError(
+            TransactionRequestError::MissingAuthenticatedInputNote(_)
+        )
+    ));
+}
+
+#[tokio::test]
+async fn swap_chain_test() {
+    // This test simulates a "swap chain" scenario with multiple wallets and fungible assets.
+    // 1. It creates a number wallet/faucet pairs, each wallet holding an asset minted by its paired
+    //    faucet.
+    // 2. For each consecutive pair, it creates a swap transaction where wallet N offers its asset
+    //    and requests the asset of wallet N+1.
+    // 3. The last wallet, which didn't generate any swaps, holds the asset that the wallet N-1
+    //    requested, which in turn was the asset requested by wallet N-2, and so on.
+    // 4. The test then consumes all swap notes (in reverse order) in a single transaction against
+    //    the last wallet.
+    // 5. Although the last wallet doesn't contain any of the intermediate requested assets, it
+    //    should be able to consume the swap notes because it will hold the requested asset for each
+    //    step and gain the needed asset for the next. This can only happen if the notes are
+    //    consumed in the specified order.
+    // 6. Finally, it asserts that the last wallet now owns the asset originally held by the first
+    //    wallet, verifying that the whole swap chain was successful.
+
+    let (mut client, _, keystore) = create_test_client().await;
+
+    // Generate a few account pairs with a fungible asset that can be used for swaps.
+    let (wallets, faucets) = setup_accounts_and_faucets(
+        &mut client,
+        &keystore,
+        AccountStorageMode::Private,
+        5, // Number of wallets
+        5, // Number of faucets
+        vec![
+            // Each wallet will have a different asset
+            vec![1, 0, 0, 0, 0],
+            vec![0, 1, 0, 0, 0],
+            vec![0, 0, 1, 0, 0],
+            vec![0, 0, 0, 1, 0],
+            vec![0, 0, 0, 0, 1],
+        ],
+    )
+    .await
+    .unwrap();
+
+    // Generate swap notes.
+    // Execpt for the last, each wallet N will offer it's faucet N asset and request a faucet N+1
+    // asset.
+    let account_pairs: Vec<_> = wallets.into_iter().zip(faucets.into_iter()).collect();
+    let mut swap_notes = vec![];
+    for pairs in account_pairs.windows(2) {
+        let tx_request = TransactionRequestBuilder::new()
+            .build_swap(
+                &SwapTransactionData::new(
+                    pairs[0].0.id(),
+                    Asset::Fungible(FungibleAsset::new(pairs[0].1.id(), 1).unwrap()),
+                    Asset::Fungible(FungibleAsset::new(pairs[1].1.id(), 1).unwrap()),
+                ),
+                NoteType::Private,
+                client.rng(),
+            )
+            .unwrap();
+
+        // The notes are inserted in reverse order because the first note to be consumed will be the
+        // last one generated.
+        swap_notes.insert(0, tx_request.expected_output_own_notes()[0].id());
+        execute_tx_and_sync(&mut client, pairs[0].0.id(), tx_request).await.unwrap();
+    }
+
+    // The last wallet didn't generate any swap notes and has the asset needed to start the swap
+    // chain.
+    let last_wallet = account_pairs.last().unwrap().0.id();
+
+    // Trying to consume the notes in another order will fail.
+    let tx_request = TransactionRequestBuilder::new()
+        .build_consume_notes(swap_notes.iter().rev().copied().collect())
+        .unwrap();
+    let error = client.new_transaction(last_wallet, tx_request).await.unwrap_err();
+    assert!(matches!(
+        error,
+        ClientError::TransactionExecutorError(
+            TransactionExecutorError::TransactionProgramExecutionFailed(_)
+        )
+    ));
+
+    let tx_request = TransactionRequestBuilder::new().build_consume_notes(swap_notes).unwrap();
+    execute_tx(&mut client, last_wallet, tx_request).await.unwrap();
+
+    // At the end, the last wallet should have the asset of the first wallet.
+    let last_wallet_account = client.get_account(last_wallet).await.unwrap().unwrap();
+    assert_eq!(
+        last_wallet_account
+            .account()
+            .vault()
+            .get_balance(account_pairs[0].1.id())
+            .unwrap(),
+        1
+    );
 }